refactored Gmail as a standalone class
