+++ extension/chrome/dev/export.ts
@@ -4,18 +4,20 @@
 
 import { Catch } from '../../js/common/platform/catch.js';
 import { Ui, Browser } from '../../js/common/browser.js';
-import { Google, GmailRes } from '../../js/common/api/google.js';
 import { Assert } from '../../js/common/assert.js';
 import { Api } from '../../js/common/api/api.js';
 import { Att } from '../../js/common/core/att.js';
 import { Buf } from '../../js/common/core/buf.js';
 import { openpgp } from '../../js/common/core/pgp.js';
 import { Url } from '../../js/common/core/common.js';
+import { Gmail } from '../../js/common/api/email_provider/gmail/gmail.js';
+import { GmailRes, GmailParser } from '../../js/common/api/email_provider/gmail/gmail-parser.js';
 
 Catch.try(async () => {
 
   const uncheckedUrlParams = Url.parse(['acctEmail']);
   const acctEmail = Assert.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
+  const gmail = new Gmail(acctEmail);
 
   if (!confirm('This is page is meant for debugging. It will download messages from your inbox and save them to your device. Continue?')) {
     window.close();
@@ -44,7 +46,7 @@ Catch.try(async () => {
     const msgMetas: GmailRes.GmailMsgList$message[] = [];
     let nextCyclePageToken: string | undefined;
     while (true) {
-      const { messages, resultSizeEstimate, nextPageToken } = await Google.gmail.msgList(acctEmail, 'is:inbox OR is:sent', false, nextCyclePageToken);
+      const { messages, resultSizeEstimate, nextPageToken } = await gmail.msgList('is:inbox OR is:sent', false, nextCyclePageToken);
       print(`msgList: ${(messages || []).length} msgs, resultSizeEstimate:${resultSizeEstimate}, nextPageToken: ${nextPageToken}`);
       msgMetas.push(...(messages || []));
       if (!messages || !messages.length || !nextPageToken) {
@@ -57,18 +59,18 @@ Catch.try(async () => {
     const draftMetas: GmailRes.GmailDraftMeta[] = [];
     let draftNextPageToken: string | undefined | null;
     do {
-      const { drafts, nextPageToken } = await Google.gmail.draftList(acctEmail);
+      const { drafts, nextPageToken } = await gmail.draftList();
       draftMetas.push(...drafts);
       draftNextPageToken = nextPageToken;
     } while (draftNextPageToken);
     print(`found ${draftMetas.length} drafts`);
     const fullMsgIdsList = (msgMetas || []).map(m => m.id).concat(draftMetas.map(dm => dm.message.id));
     print(`downloading full..`);
-    const msgsFull = await Google.gmail.msgsGet(acctEmail, fullMsgIdsList, 'full');
+    const msgsFull = await gmail.msgsGet(fullMsgIdsList, 'full');
     print(`downloading full done. waiting 5 seconds..`);
     await Ui.time.sleep(5000);
     print(`waiting done. Downloading raw..`);
-    const msgsRaw = await Google.gmail.msgsGet(acctEmail, fullMsgIdsList, 'raw');
+    const msgsRaw = await gmail.msgsGet(fullMsgIdsList, 'raw');
     print(`downloading raw done. Joining results..`);
     for (const msg of msgsFull) {
       for (const msgRaw of msgsRaw) {
@@ -92,14 +94,14 @@ Catch.try(async () => {
     }
     const messages: GmailRes.GmailMsg[] = [...msgsFull];
     print(`joining done. Downloading labels..`);
-    const { labels } = await Google.gmail.labelsGet(acctEmail);
+    const { labels } = await gmail.labelsGet();
     print('labels done. waiting 5s..');
     await Ui.time.sleep(5000);
     print('waiting done. Downloading attachments..');
     const fetchableAtts: Att[] = [];
     const skippedAtts: Att[] = [];
     for (const msg of messages) {
-      for (const att of Google.gmail.findAtts(msg)) {
+      for (const att of GmailParser.findAtts(msg)) {
         if (att.length > 1024 * 1024 * 7) { // over 7 mb - attachment too big
           skippedAtts.push(new Att({ data: Buf.fromUtfStr(`MOCK: ATTACHMENT STRIPPED - ORIGINAL SIZE ${att.length}`), id: att.id, msgId: msg.id }));
         } else {
@@ -107,7 +109,7 @@ Catch.try(async () => {
         }
       }
     }
-    await Google.gmail.fetchAtts(acctEmail, fetchableAtts, percent => print(`Percent atts done: ${percent}`));
+    await gmail.fetchAtts(fetchableAtts, percent => print(`Percent atts done: ${percent}`));
     const attachments: { [id: string]: { data: string, size: number } } = {};
     for (const att of fetchableAtts.concat(skippedAtts)) {
       attachments[att.id!] = { data: att.getData().toBase64UrlStr(), size: att.getData().length };
