+++ extension/chrome/elements/attachment.ts
@@ -9,10 +9,10 @@ import { Api } from '../../js/common/api/api.js';
 import { DecryptErrTypes, PgpMsg } from '../../js/common/core/pgp.js';
 import { BrowserMsg } from '../../js/common/extension.js';
 import { Att } from '../../js/common/core/att.js';
-import { Google } from '../../js/common/api/google.js';
 import { Assert } from '../../js/common/assert.js';
 import { Xss } from '../../js/common/platform/xss.js';
 import { Url } from '../../js/common/core/common.js';
+import { Gmail } from '../../js/common/api/email_provider/gmail/gmail.js';
 
 Catch.try(async () => {
 
@@ -35,6 +35,7 @@ Catch.try(async () => {
   const button = $('#download');
   let origHtmlContent: string;
   let progressEl: JQuery<HTMLElement>;
+  const gmail = new Gmail(acctEmail);
 
   let att: Att;
   try {
@@ -139,7 +140,7 @@ Catch.try(async () => {
     if (a.url) { // when content was downloaded and decrypted
       a.setData(await Api.download(a.url, renderProgress));
     } else if (a.id && a.msgId) { // gmail attId
-      const { data } = await Google.gmail.attGet(acctEmail, a.msgId, a.id, renderProgress);
+      const { data } = await gmail.attGet(a.msgId, a.id, renderProgress);
       a.setData(data);
     } else {
       throw new Error('File is missing both id and url - this should be fixed');
@@ -166,7 +167,7 @@ Catch.try(async () => {
 
   const recoverMissingAttIdIfNeeded = async (a: Att) => {
     if (!a.url && !a.id && a.msgId) {
-      const result = await Google.gmail.msgGet(acctEmail, a.msgId, 'full');
+      const result = await gmail.msgGet(a.msgId, 'full');
       if (result && result.payload && result.payload.parts) {
         for (const attMeta of result.payload.parts) {
           if (attMeta.filename === name && attMeta.body && attMeta.body.size === size && attMeta.body.attachmentId) {
@@ -182,7 +183,7 @@ Catch.try(async () => {
 
   const processAsPublicKeyAndHideAttIfAppropriate = async (a: Att) => {
     if (a.msgId && a.id && a.treatAs() === 'publicKey') { // this is encrypted public key - download && decrypt & parse & render
-      const { data } = await Google.gmail.attGet(acctEmail, a.msgId, a.id);
+      const { data } = await gmail.attGet(a.msgId, a.id);
       const decrRes = await PgpMsg.decrypt({ kisWithPp: await Store.keysGetAllWithPp(acctEmail), encryptedData: data });
       if (decrRes.success && decrRes.content) {
         const openpgpType = await PgpMsg.type({ data: decrRes.content });
