+++ extension/js/common/api/google.ts
@@ -8,90 +8,24 @@ const BUILD = 'consumer'; // todo
 
 import { Catch } from '../platform/catch.js';
 import { Store, AccountStore, Serializable } from '../platform/store.js';
-import { Api, ReqMethod, ProgressCbs, ProgressCb, AjaxErr, RecipientType, ChunkedCb, ProviderContactsResults, GoogleAuthErr } from './api.js';
-import { Env, Ui } from '../browser.js';
-import { Dict, Value, Str, Url } from '../core/common.js';
-import { GoogleAuthWindowResult$result, BrowserMsg, AddrParserResult, BrowserWidnow } from '../extension.js';
-import { Mime, SendableMsgBody } from '../core/mime.js';
-import { Att } from '../core/att.js';
-import { FormatError, Pgp, Contact } from '../core/pgp.js';
+import { Api, ReqMethod, ProgressCbs, GoogleAuthErr } from './api.js';
+import { Ui } from '../browser.js';
+import { Dict, Value, Url, Str } from '../core/common.js';
+import { GoogleAuthWindowResult$result } from '../extension.js';
 import { tabsQuery, windowsCreate } from './chrome.js';
 import { Buf } from '../core/buf.js';
-import { gmailBackupSearchQuery, GOOGLE_API_HOST, GOOGLE_OAUTH_SCREEN_HOST } from '../core/const.js';
-import { EmailProviderApi, SendableMsg } from './email_provider_api.js';
-import { Xss } from '../platform/xss.js';
+import { GOOGLE_API_HOST, GOOGLE_OAUTH_SCREEN_HOST } from '../core/const.js';
+import { EmailProviderApi } from './email_provider/email_provider_api.js';
+import { GmailRes } from './email_provider/gmail/gmail-parser.js';
 
 type GoogleAuthTokenInfo = { issued_to: string, audience: string, scope: string, expires_in: number, access_type: 'offline' };
 type GoogleAuthTokensResponse = { access_token: string, expires_in: number, refresh_token?: string, id_token: string, token_type: 'Bearer' };
 export type AuthReq = { acctEmail?: string, scopes: string[], messageId?: string, csrfToken: string };
-export type GmailResponseFormat = 'raw' | 'full' | 'metadata';
 type AuthResultSuccess = { result: 'Success', acctEmail: string, id_token: string, error?: undefined };
 type AuthResultError = { result: GoogleAuthWindowResult$result, acctEmail?: string, error?: string, id_token: undefined };
 export type AuthRes = AuthResultSuccess | AuthResultError;
 
-export namespace GmailRes { // responses
-
-  export type GmailMsg$header = { name: string, value: string };
-  export type GmailMsg$payload$body = { attachmentId: string, size: number, data?: string };
-  export type GmailMsg$payload$part = { body?: GmailMsg$payload$body, filename?: string, mimeType?: string, headers?: GmailMsg$header[] };
-  export type GmailMsg$payload = { parts?: GmailMsg$payload$part[], headers?: GmailMsg$header[], mimeType?: string, body?: GmailMsg$payload$body };
-  export type GmailMsg$labelId = 'INBOX' | 'UNREAD' | 'CATEGORY_PERSONAL' | 'IMPORTANT' | 'SENT' | 'CATEGORY_UPDATES' | 'TRASH';
-  export type GmailMsg = {
-    id: string; historyId: string; threadId?: string | null; payload?: GmailMsg$payload; internalDate?: number | string;
-    labelIds?: GmailMsg$labelId[]; snippet?: string; raw?: string;
-  };
-  export type GmailMsgList$message = { id: string, threadId: string };
-  export type GmailMsgList = { messages?: GmailMsgList$message[], resultSizeEstimate: number, nextPageToken?: string };
-  export type GmailLabels$label = {
-    id: string, name: string, messageListVisibility: 'show' | 'hide', labelListVisibility: 'labelShow' | 'labelHide', type: 'user' | 'system',
-    messagesTotal?: number, messagesUnread?: number, threadsTotal?: number, threadsUnread?: number, color?: { textColor: string, backgroundColor: string }
-  };
-  export type GmailLabels = { labels: GmailLabels$label[] };
-  export type GmailAtt = { attachmentId: string, size: number, data: Buf };
-  export type GmailMsgSend = { id: string };
-  export type GmailThread = { id: string, historyId: string, messages: GmailMsg[] };
-  export type GmailThreadList = { threads: { historyId: string, id: string, snippet: string }[], nextPageToken: string, resultSizeEstimate: number };
-  export type GmailDraftCreate = { id: string };
-  export type GmailDraftDelete = {};
-  export type GmailDraftUpdate = {};
-  export type GmailDraftGet = { id: string, message: GmailMsg };
-  export type GmailDraftMeta = { id: string, message: { id: string, threadId: string } };
-  export type GmailDraftList = { drafts: GmailDraftMeta[], nextPageToken: string };
-  export type GmailDraftSend = {};
-  export type GmailAliases = { sendAs: GmailAliases$sendAs[] };
-  type GmailAliases$sendAs = {
-    sendAsEmail: string, displayName: string, replyToAddress: string, signature: string,
-    isDefault: boolean, treatAsAlias: boolean, verificationStatus: string, isPrimary?: true
-  };
-
-  export type OpenId = { // 'name' is the full name, picture is url
-    at_hash: string; exp: number; iat: number; sub: string; aud: string; azp: string; iss: "https://accounts.google.com";
-    name: string; picture: string; locale: 'en' | string; family_name: string; given_name: string;
-    email?: string, email_verified?: boolean;
-  };
-
-  export type GoogleContacts = {
-    feed: {
-      entry?: {
-        gd$email?: {
-          address: string,
-          primary: string
-        }[],
-        gd$name?: {
-          gd$fullName?: {
-            $t: string
-          }
-        }
-      }[]
-    }
-  };
-
-}
-
 export class Google extends EmailProviderApi {
-
-  private static GMAIL_USELESS_CONTACTS_FILTER = '-to:txt.voice.google.com -to:craigslist.org';
-  private static GMAIL_SEARCH_QUERY_LENGTH_LIMIT = 1400;
 
   public static webmailUrl = (acctEmail: string) => `https://mail.google.com/mail/u/${acctEmail}`;
 
@@ -134,7 +68,7 @@ export class Google extends EmailProviderApi {
       }));
   }
 
-  private static encodeAsMultipartRelated = (parts: Dict<string>) => { // todo - this could probably be achieved with emailjs-mime-builder
+  static encodeAsMultipartRelated(parts: Dict<string>) { // todo - this could probably be achieved with emailjs-mime-builder
     const boundary = 'the_boundary_is_' + Str.sloppyRandom(10);
     let body = '';
     for (const type of Object.keys(parts)) {
@@ -151,427 +85,7 @@ export class Google extends EmailProviderApi {
     return { contentType: 'multipart/related; boundary=' + boundary, body };
   }
 
-  public static gmail = {
-    buildSearchQuery: {
-      or: (arr: string[], quoted: boolean = false) => {
-        if (quoted) {
-          return '("' + arr.join('") OR ("') + '")';
-        } else {
-          return '(' + arr.join(') OR (') + ')';
-        }
-      },
-    },
-    threadGet: (acctEmail: string, threadId: string, format?: GmailResponseFormat, progressCb?: ProgressCb): Promise<GmailRes.GmailThread> =>
-      Google.gmailCall(acctEmail, 'GET', `threads/${threadId}`, {
-        format,
-      }, { download: progressCb }),
-    threadList: (acctEmail: string, labelId: string): Promise<GmailRes.GmailThreadList> => Google.gmailCall(acctEmail, 'GET', `threads`, {
-      labelIds: labelId !== 'ALL' ? labelId : undefined,
-      includeSpamTrash: Boolean(labelId === 'SPAM' || labelId === 'TRASH'),
-      // pageToken: page_token,
-      // q,
-      // maxResults
-    }),
-    threadModify: (acctEmail: string, id: string, rmLabels: string[], addLabels: string[]): Promise<GmailRes.GmailThread> => Google.gmailCall(acctEmail, 'POST', `threads/${id}/modify`, {
-      removeLabelIds: rmLabels || [], // todo - insufficient permission - need https://github.com/FlowCrypt/flowcrypt-browser/issues/1304
-      addLabelIds: addLabels || [],
-    }),
-    draftCreate: (acctEmail: string, mimeMsg: string, threadId: string): Promise<GmailRes.GmailDraftCreate> => Google.gmailCall(acctEmail, 'POST', 'drafts', {
-      message: {
-        raw: Buf.fromUtfStr(mimeMsg).toBase64UrlStr(),
-        threadId,
-      },
-    }),
-    draftDelete: (acctEmail: string, id: string): Promise<GmailRes.GmailDraftDelete> => Google.gmailCall(acctEmail, 'DELETE', 'drafts/' + id, undefined),
-    draftUpdate: (acctEmail: string, id: string, mimeMsg: string): Promise<GmailRes.GmailDraftUpdate> => Google.gmailCall(acctEmail, 'PUT', `drafts/${id}`, {
-      message: {
-        raw: Buf.fromUtfStr(mimeMsg).toBase64UrlStr(),
-      },
-    }),
-    draftGet: (acctEmail: string, id: string, format: GmailResponseFormat = 'full'): Promise<GmailRes.GmailDraftGet> => Google.gmailCall(acctEmail, 'GET', `drafts/${id}`, {
-      format,
-    }),
-    draftList: (acctEmail: string): Promise<GmailRes.GmailDraftList> => Google.gmailCall(acctEmail, 'GET', 'drafts', undefined),
-    draftSend: (acctEmail: string, id: string): Promise<GmailRes.GmailDraftSend> => Google.gmailCall(acctEmail, 'POST', 'drafts/send', {
-      id,
-    }),
-    msgSend: async (acctEmail: string, message: SendableMsg, progressCb?: ProgressCb): Promise<GmailRes.GmailMsgSend> => {
-      message.headers.From = message.from;
-      for (const key of Object.keys(message.recipients)) {
-        const sendingType = key as RecipientType;
-        if (message.recipients[sendingType] && message.recipients[sendingType]!.length) {
-          // todo - properly escape/encode this header using emailjs
-          message.headers[sendingType[0].toUpperCase() + sendingType.slice(1)] = message.recipients[sendingType]!.map(h => h.replace(/[,]/g, '')).join(',');
-        }
-      }
-      message.headers.Subject = message.subject;
-      const mimeMsg = await Mime.encode(message.body, message.headers, message.atts, message.mimeRootType, message.sign);
-      const request = Google.encodeAsMultipartRelated({ 'application/json; charset=UTF-8': JSON.stringify({ threadId: message.thread }), 'message/rfc822': mimeMsg });
-      return Google.gmailCall(acctEmail, 'POST', 'messages/send', request.body, { upload: progressCb || Value.noop }, request.contentType);
-    },
-    msgList: (acctEmail: string, q: string, includeDeleted: boolean = false, pageToken?: string): Promise<GmailRes.GmailMsgList> => Google.gmailCall(acctEmail, 'GET', 'messages', {
-      q,
-      includeSpamTrash: includeDeleted,
-      pageToken,
-    }),
-    /**
-     * Attempting to `msgGet format:raw` from within content scripts would likely fail if the mime message is 1MB or larger,
-     * because strings over 1 MB may fail to get to/from bg page. A way to mitigate that would be to pass `R.GmailMsg$raw` prop
-     * as a Buf instead of a string.
-    */
-    msgGet: async (acctEmail: string, msgId: string, format: GmailResponseFormat, progressCb?: ProgressCb): Promise<GmailRes.GmailMsg> =>
-      Google.gmailCall(acctEmail, 'GET', `messages/${msgId}`, {
-        format: format || 'full'
-      }, { download: progressCb }),
-    msgsGet: (acctEmail: string, msgIds: string[], format: GmailResponseFormat): Promise<GmailRes.GmailMsg[]> => {
-      return Promise.all(msgIds.map(id => Google.gmail.msgGet(acctEmail, id, format)));
-    },
-    labelsGet: (acctEmail: string): Promise<GmailRes.GmailLabels> => Google.gmailCall(acctEmail, 'GET', `labels`, {}),
-    attGet: async (acctEmail: string, msgId: string, attId: string, progressCb?: ProgressCb): Promise<GmailRes.GmailAtt> => {
-      type RawGmailAttRes = { attachmentId: string, size: number, data: string };
-      const { attachmentId, size, data } = await Google.gmailCall(acctEmail, 'GET', `messages/${msgId}/attachments/${attId}`, {}, { download: progressCb }) as RawGmailAttRes;
-      return { attachmentId, size, data: Buf.fromBase64UrlStr(data) }; // data should be a Buf for ease of passing to/from bg page
-    },
-    attGetChunk: (acctEmail: string, msgId: string, attId: string): Promise<Buf> => new Promise((resolve, reject) => {
-      if (Env.isContentScript()) {
-        // content script CORS not allowed anymore, have to drag it through background page
-        // https://www.chromestatus.com/feature/5629709824032768
-        BrowserMsg.send.bg.await.ajaxGmailAttGetChunk({ acctEmail, msgId, attId }).then(({ chunk }) => resolve(chunk)).catch(reject);
-        return;
-      }
-      const stack = Catch.stackTrace();
-      const minBytes = 1000;
-      let processed = 0;
-      const processChunkAndResolve = (chunk: string) => {
-        if (!processed++) {
-          // make json end guessing easier
-          chunk = chunk.replace(/[\n\s\r]/g, '');
-          // the response is a chunk of json that may not have ended. One of:
-          // {"length":123,"data":"kks
-          // {"length":123,"data":"kksdwei
-          // {"length":123,"data":"kksdwei"
-          // {"length":123,"data":"kksdwei"}
-          if (chunk[chunk.length - 1] !== '"' && chunk[chunk.length - 2] !== '"') {
-            chunk += '"}'; // json end
-          } else if (chunk[chunk.length - 1] !== '}') {
-            chunk += '}'; // json end
-          }
-          let parsedJsonDataField;
-          try {
-            parsedJsonDataField = JSON.parse(chunk).data; // tslint:disable-line:no-unsafe-any
-          } catch (e) {
-            console.info(e);
-            reject(new Error("Chunk response could not be parsed"));
-            return;
-          }
-          for (let i = 0; parsedJsonDataField && i < 50; i++) {
-            try {
-              resolve(Buf.fromBase64UrlStr(parsedJsonDataField)); // tslint:disable-line:no-unsafe-any
-              return;
-            } catch (e) {
-              // the chunk of data may have been cut at an inconvenient index
-              // shave off up to 50 trailing characters until it can be decoded
-              parsedJsonDataField = parsedJsonDataField.slice(0, -1); // tslint:disable-line:no-unsafe-any
-            }
-          }
-          reject(new Error("Chunk response could not be decoded"));
-        }
-      };
-      GoogleAuth.googleApiAuthHeader(acctEmail).then(authToken => {
-        const r = new XMLHttpRequest();
-        const method = 'GET';
-        const url = `${GOOGLE_API_HOST}/gmail/v1/users/me/messages/${msgId}/attachments/${attId}`;
-        r.open(method, url, true);
-        r.setRequestHeader('Authorization', authToken);
-        r.send();
-        let status: number;
-        const responsePollInterval = Catch.setHandledInterval(() => {
-          if (status >= 200 && status <= 299 && r.responseText.length >= minBytes) {
-            window.clearInterval(responsePollInterval);
-            processChunkAndResolve(r.responseText);
-            r.abort();
-          }
-        }, 10);
-        r.onreadystatechange = () => {
-          if (r.readyState === 2 || r.readyState === 3) { // headers, loading
-            status = r.status;
-            if (status >= 300) {
-              reject(AjaxErr.fromXhr({ status, readyState: r.readyState }, { method, url }, stack));
-              window.clearInterval(responsePollInterval);
-              r.abort();
-            }
-          }
-          if (r.readyState === 3 || r.readyState === 4) { // loading, done
-            if (status >= 200 && status <= 299 && r.responseText.length >= minBytes) { // done as a success - resolve in case response_poll didn't catch this yet
-              processChunkAndResolve(r.responseText);
-              window.clearInterval(responsePollInterval);
-              if (r.readyState === 3) {
-                r.abort();
-              }
-            } else { // done as a fail - reject
-              reject(AjaxErr.fromXhr({ status, readyState: r.readyState }, { method, url }, stack));
-              window.clearInterval(responsePollInterval);
-            }
-          }
-        };
-      }).catch(reject);
-    }),
-    findHeader: (apiGmailMsgObj: GmailRes.GmailMsg | GmailRes.GmailMsg$payload, headerName: string) => {
-      const node: GmailRes.GmailMsg$payload = apiGmailMsgObj.hasOwnProperty('payload') ? (apiGmailMsgObj as GmailRes.GmailMsg).payload! : apiGmailMsgObj as GmailRes.GmailMsg$payload;
-      if (typeof node.headers !== 'undefined') {
-        for (const header of node.headers) {
-          if (header.name.toLowerCase() === headerName.toLowerCase()) {
-            return header.value;
-          }
-        }
-      }
-      return undefined;
-    },
-    findAtts: (msgOrPayloadOrPart: GmailRes.GmailMsg | GmailRes.GmailMsg$payload | GmailRes.GmailMsg$payload$part, internalResults: Att[] = [], internalMsgId?: string) => {
-      if (msgOrPayloadOrPart.hasOwnProperty('payload')) {
-        internalMsgId = (msgOrPayloadOrPart as GmailRes.GmailMsg).id;
-        Google.gmail.findAtts((msgOrPayloadOrPart as GmailRes.GmailMsg).payload!, internalResults, internalMsgId);
-      }
-      if (msgOrPayloadOrPart.hasOwnProperty('parts')) {
-        for (const part of (msgOrPayloadOrPart as GmailRes.GmailMsg$payload).parts!) {
-          Google.gmail.findAtts(part, internalResults, internalMsgId);
-        }
-      }
-      if (msgOrPayloadOrPart.hasOwnProperty('body') && (msgOrPayloadOrPart as GmailRes.GmailMsg$payload$part).body!.hasOwnProperty('attachmentId')) {
-        internalResults.push(new Att({
-          msgId: internalMsgId,
-          id: (msgOrPayloadOrPart as GmailRes.GmailMsg$payload$part).body!.attachmentId,
-          length: (msgOrPayloadOrPart as GmailRes.GmailMsg$payload$part).body!.size,
-          name: (msgOrPayloadOrPart as GmailRes.GmailMsg$payload$part).filename,
-          type: (msgOrPayloadOrPart as GmailRes.GmailMsg$payload$part).mimeType,
-          inline: (Google.gmail.findHeader(msgOrPayloadOrPart, 'content-disposition') || '').toLowerCase().indexOf('inline') === 0,
-        }));
-      }
-      return internalResults;
-    },
-    findBodies: (gmailMsg: GmailRes.GmailMsg | GmailRes.GmailMsg$payload | GmailRes.GmailMsg$payload$part, internalResults: SendableMsgBody = {}): SendableMsgBody => {
-      const isGmailMsgWithPayload = (v: any): v is GmailRes.GmailMsg => v && typeof (v as GmailRes.GmailMsg).payload !== 'undefined';
-      const isGmailMsgPayload = (v: any): v is GmailRes.GmailMsg$payload => v && typeof (v as GmailRes.GmailMsg$payload).parts !== 'undefined';
-      const isGmailMsgPayloadPart = (v: any): v is GmailRes.GmailMsg$payload$part => v && typeof (v as GmailRes.GmailMsg$payload$part).body !== 'undefined';
-      if (isGmailMsgWithPayload(gmailMsg)) {
-        Google.gmail.findBodies(gmailMsg.payload!, internalResults);
-      }
-      if (isGmailMsgPayload(gmailMsg) && gmailMsg.parts) {
-        for (const part of gmailMsg.parts) {
-          Google.gmail.findBodies(part, internalResults);
-        }
-      }
-      if (isGmailMsgPayloadPart(gmailMsg) && gmailMsg.body && typeof gmailMsg.body.data !== 'undefined' && gmailMsg.body.size !== 0) {
-        if (gmailMsg.mimeType) {
-          internalResults[gmailMsg.mimeType] = gmailMsg.body.data;
-        }
-      }
-      return internalResults;
-    },
-    fetchAtts: async (acctEmail: string, atts: Att[], progressCb?: ProgressCb) => {
-      if (!atts.length) {
-        return;
-      }
-      let lastProgressPercent = -1;
-      const loadedAr: Array<number> = [];
-      // 1.33 is a coefficient we need to multiply because total size we need to download is larger than all files together
-      const total = atts.map(x => x.length).reduce((a, b) => a + b) * 1.33;
-      const responses = await Promise.all(atts.map((a, index) => Google.gmail.attGet(acctEmail, a.msgId!, a.id!, (_, loaded, s) => {
-        if (progressCb) {
-          loadedAr[index] = loaded || 0;
-          const totalLoaded = loadedAr.reduce((a, b) => a + b);
-          const progressPercent = Math.round((totalLoaded * 100) / total);
-          if (progressPercent !== lastProgressPercent) {
-            lastProgressPercent = progressPercent;
-            progressCb(progressPercent, totalLoaded, total);
-          }
-        }
-      })));
-      for (const i of responses.keys()) {
-        atts[i].setData(responses[i].data);
-      }
-    },
-    /**
-    * This will keep triggering callback with new emails as they are being discovered
-    */
-    searchContacts: async (acctEmail: string, userQuery: string, knownContacts: Contact[], chunkedCb: ChunkedCb) => {
-      let gmailQuery = `is:sent ${Google.GMAIL_USELESS_CONTACTS_FILTER} `;
-      if (userQuery) {
-        const variationsOfTo = userQuery.split(/[ .]/g).filter(v => !['com', 'org', 'net'].includes(v));
-        if (!variationsOfTo.includes(userQuery)) {
-          variationsOfTo.push(userQuery);
-        }
-        gmailQuery += '(';
-        while (variationsOfTo.length) {
-          gmailQuery += `to:${variationsOfTo.pop()}`;
-          if (gmailQuery.length > Google.GMAIL_SEARCH_QUERY_LENGTH_LIMIT) {
-            break;
-          }
-          if (variationsOfTo.length > 1) {
-            gmailQuery += ' OR ';
-          }
-        }
-        gmailQuery += ')';
-      }
-      for (const contact of knownContacts.filter(c => Str.isEmailValid(c.email))) {
-        if (gmailQuery.length > Google.GMAIL_SEARCH_QUERY_LENGTH_LIMIT) {
-          break;
-        }
-        gmailQuery += ` -to:${contact.email}`;
-      }
-      await Google.apiGmailLoopThroughEmailsToCompileContacts(acctEmail, gmailQuery, chunkedCb);
-    },
-    /**
-     * Extracts the encrypted message from gmail api. Sometimes it's sent as a text, sometimes html, sometimes attachments in various forms.
-     */
-    extractArmoredBlock: async (acctEmail: string, msgId: string, format: GmailResponseFormat, progressCb?: ProgressCb): Promise<{ armored: string, subject?: string }> => {
-      const gmailMsg = await Google.gmail.msgGet(acctEmail, msgId, format);
-      const subject = gmailMsg.payload ? Google.gmail.findHeader(gmailMsg.payload, 'subject') : undefined;
-      if (format === 'full') {
-        const bodies = Google.gmail.findBodies(gmailMsg);
-        const atts = Google.gmail.findAtts(gmailMsg);
-        const fromTextBody = Pgp.armor.clip(Buf.fromBase64UrlStr(bodies['text/plain'] || '').toUtfStr());
-        if (fromTextBody) {
-          return { armored: fromTextBody, subject };
-        }
-        const fromHtmlBody = Pgp.armor.clip(Xss.htmlSanitizeAndStripAllTags(Buf.fromBase64UrlStr(bodies['text/html'] || '').toUtfStr(), '\n'));
-        if (fromHtmlBody) {
-          return { armored: fromHtmlBody, subject };
-        }
-        if (atts.length) {
-          for (const att of atts) {
-            if (att.treatAs() === 'encryptedMsg') {
-              await Google.gmail.fetchAtts(acctEmail, [att], progressCb);
-              const armoredMsg = Pgp.armor.clip(att.getData().toUtfStr());
-              if (!armoredMsg) {
-                throw new FormatError('Problem extracting armored message', att.getData().toUtfStr());
-              }
-              return { armored: armoredMsg, subject };
-            }
-          }
-          throw new FormatError('Armored message not found', JSON.stringify(gmailMsg.payload, undefined, 2));
-        } else {
-          throw new FormatError('No attachments', JSON.stringify(gmailMsg.payload, undefined, 2));
-        }
-      } else { // format === raw
-        const mimeMsg = Buf.fromBase64UrlStr(gmailMsg.raw!);
-        const decoded = await Mime.decode(mimeMsg);
-        if (decoded.text !== undefined) {
-          const armoredMsg = Pgp.armor.clip(decoded.text); // todo - the message might be in attachments
-          if (armoredMsg) {
-            return { armored: armoredMsg, subject };
-          } else {
-            throw new FormatError('Could not find armored message in parsed raw mime', mimeMsg.toUtfStr());
-          }
-        } else {
-          throw new FormatError('No text in parsed raw mime', mimeMsg.toUtfStr());
-        }
-      }
-    },
-    fetchAcctAliases: async (acctEmail: string): Promise<GmailRes.GmailAliases> => Google.gmailCall(acctEmail, 'GET', 'settings/sendAs', {}),
-    fetchMsgsHeadersBasedOnQuery: async (acctEmail: string, q: string, headerNames: string[], msgLimit: number) => {
-      const { messages } = await Google.gmail.msgList(acctEmail, q, false);
-      return await Google.extractHeadersFromMsgs(acctEmail, messages || [], headerNames, msgLimit);
-    },
-    fetchKeyBackups: async (acctEmail: string) => {
-      const res = await Google.gmail.msgList(acctEmail, gmailBackupSearchQuery(acctEmail), true);
-      if (!res.messages) {
-        return [];
-      }
-      const msgIds = res.messages.map(m => m.id);
-      const msgs = await Google.gmail.msgsGet(acctEmail, msgIds, 'full');
-      const atts: Att[] = [];
-      for (const msg of msgs) {
-        atts.push(...Google.gmail.findAtts(msg));
-      }
-      await Google.gmail.fetchAtts(acctEmail, atts);
-      const { keys } = await Pgp.key.readMany(Buf.fromUtfStr(atts.map(a => a.getData().toUtfStr()).join('\n')));
-      return keys;
-    },
-  };
 
-  private static apiGmailBuildFilteredQuery = (query: string, allRawEmails: string[]) => {
-    let filteredQuery = query;
-    for (const rawEmail of allRawEmails) {
-      filteredQuery += ` -to:"${rawEmail}"`;
-      if (filteredQuery.length > Google.GMAIL_SEARCH_QUERY_LENGTH_LIMIT) {
-        return filteredQuery;
-      }
-    }
-    return filteredQuery;
-  }
-
-  private static apiGmailGetNewUniqueRecipientsFromHeaders = async (toHeaders: string[], allResults: Contact[], allRawEmails: string[]): Promise<Contact[]> => {
-    if (!toHeaders.length) {
-      return [];
-    }
-    const rawParsedResults: AddrParserResult[] = [];
-    toHeaders = Value.arr.unique(toHeaders);
-    for (const to of toHeaders) {
-      rawParsedResults.push(...(window as unknown as BrowserWidnow)['emailjs-addressparser'].parse(to));
-    }
-    for (const rawParsedRes of rawParsedResults) {
-      if (rawParsedRes.address && allRawEmails.indexOf(rawParsedRes.address) === -1) {
-        allRawEmails.push(rawParsedRes.address);
-      }
-    }
-    const newValidResults = await Promise.all(rawParsedResults
-      .filter(r => r.address && Str.isEmailValid(r.address))
-      .map(({ address, name }) => Store.dbContactObj({ email: address!, name }))); // address! because we .filter based on r.address being truthy
-    const uniqueNewValidResults: Contact[] = [];
-    for (const newValidRes of newValidResults) {
-      if (allResults.map(c => c.email).indexOf(newValidRes.email) === -1) {
-        const foundIndex = uniqueNewValidResults.map(c => c.email).indexOf(newValidRes.email);
-        if (foundIndex === -1) {
-          uniqueNewValidResults.push(newValidRes);
-        } else if (newValidRes.name && !uniqueNewValidResults[foundIndex].name) {
-          uniqueNewValidResults[foundIndex].name = newValidRes.name; // prefer to also save name if first encountered result is missing it
-        }
-      }
-    }
-    return uniqueNewValidResults;
-  }
-
-  private static apiGmailLoopThroughEmailsToCompileContacts = async (acctEmail: string, query: string, chunkedCb: (r: ProviderContactsResults) => void) => {
-    const allResults: Contact[] = [];
-    const allRawEmails: string[] = [];
-    let lastFilteredQuery = '';
-    let continueSearching = true;
-    while (continueSearching) {
-      const filteredQuery = Google.apiGmailBuildFilteredQuery(query, allRawEmails);
-      if (filteredQuery === lastFilteredQuery) {
-        break;
-      }
-      if (filteredQuery.length > Google.GMAIL_SEARCH_QUERY_LENGTH_LIMIT) {
-        continueSearching = false;
-      }
-      const headers = await Google.gmail.fetchMsgsHeadersBasedOnQuery(acctEmail, filteredQuery, ['to'], 50);
-      lastFilteredQuery = filteredQuery;
-      const uniqueNewValidResults = await Google.apiGmailGetNewUniqueRecipientsFromHeaders(headers.to, allResults, allRawEmails);
-      if (!uniqueNewValidResults.length) {
-        break;
-      }
-      allResults.push(...uniqueNewValidResults);
-      chunkedCb({ new: uniqueNewValidResults, all: allResults });
-    }
-    chunkedCb({ new: [], all: allResults });
-  }
-
-  private static extractHeadersFromMsgs = async (acctEmail: string, msgsIds: GmailRes.GmailMsgList$message[], headerNames: string[], msgLimit: number): Promise<Dict<string[]>> => {
-    const headerVals: Dict<string[]> = {};
-    for (const headerName of headerNames) {
-      headerVals[headerName] = [];
-    }
-    for (const msg of await Google.gmail.msgsGet(acctEmail, msgsIds.slice(0, msgLimit).map(m => m.id), 'metadata')) {
-      for (const headerName of headerNames) {
-        const value = Google.gmail.findHeader(msg, headerName);
-        if (typeof value !== 'undefined') {
-          headerVals[headerName].push(value);
-        }
-      }
-    }
-    return headerVals;
-  }
 }
 
 export class GoogleAuth {
