+++ extension/js/common/composer/formatters/base-mail-formatter.ts
@@ -3,7 +3,7 @@
 'use strict';
 
 import { NewMsgData } from '../interfaces/composer-types.js';
-import { SendableMsg } from '../../api/email_provider_api.js';
+import { SendableMsg } from '../../api/email_provider/email_provider_api.js';
 import { Composer } from '../composer.js';
 
 export interface MailFormatterInterface {
