+++ extension/js/common/composer/formatters/signed-msg-mail-formatter.ts
@@ -3,10 +3,9 @@
 'use strict';
 
 import { NewMsgData } from '../interfaces/composer-types.js';
-import { SendableMsg } from '../../api/email_provider_api.js';
+import { SendableMsg } from '../../api/email_provider/email_provider_api.js';
 import { PgpMsg } from '../../core/pgp.js';
 import { BrowserWidnow } from '../../extension.js';
-import { Google } from '../../api/google.js';
 import { Catch } from '../../platform/catch.js';
 import { BaseMailFormatter, MailFormatterInterface } from './base-mail-formatter.js';
 import { SendableMsgBody } from '../../core/mime.js';
@@ -33,13 +32,13 @@ export class SignedMsgMailFormatter extends BaseMailFormatter implements MailFor
       const allContacts = [...newMsg.recipients.to || [], ...newMsg.recipients.cc || [], ...newMsg.recipients.bcc || []];
       Store.dbContactUpdate(undefined, allContacts, { last_use: Date.now() }).catch(Catch.reportErr);
       const body = { 'text/plain': signedData };
-      return await Google.createMsgObj(this.acctEmail, newMsg.sender, newMsg.recipients, newMsg.subject, body, atts, this.composer.view.threadId);
+      return await this.composer.emailProvider.createMsgObj(newMsg.sender, newMsg.recipients, newMsg.subject, body, atts, this.composer.view.threadId);
     }
     // pgp/mime detached signature - it must be signed later, while being mime-encoded
     // prepare a sign function first, which will be used by Mime.encode later
     const sign = (signable: string) => PgpMsg.sign(signingPrv, signable, true);
     const body: SendableMsgBody = { 'text/plain': newMsg.plaintext, 'text/html': newMsg.plainhtml };
-    return await Google.createMsgObj(this.acctEmail, newMsg.sender, newMsg.recipients, newMsg.subject, body, atts, this.composer.view.threadId, undefined, sign);
+    return await this.composer.emailProvider.createMsgObj(newMsg.sender, newMsg.recipients, newMsg.subject, body, atts, this.composer.view.threadId, undefined, sign);
   }
 
 }
